spp_vf: fix build
build error:
...spp/src/vf/classifier_mac.c:314:5:
   error: format specifies type 'unsigned short' but the argument
          has type 'int' [-Werror,-Wformat]
	classified_data->num_pkt - n_tx, classified_data->port);
	^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
...build/include/rte_log.h:316:32:
   note: expanded from macro 'RTE_LOG'
	 RTE_LOGTYPE_ ## t, # t ": " __VA_ARGS__)
				     ^~~~~~~~~~~

Looks like caused by integer promotion, to fix cast variable
explicityly to uint16_t.

Fixes: 11512d8b6c28 ("spp_vf: add vf functions")

Acked-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>

+++ src/vf/classifier_mac.c
@@ -311,7 +311,8 @@ transmit_packet(struct classified_data *classified_data)
 			rte_pktmbuf_free(classified_data->pkts[i]);
 		RTE_LOG(DEBUG, SPP_CLASSIFIER_MAC,
 				"drop packets(tx). num=%hu, dpdk_port=%hu\n",
-				classified_data->num_pkt - n_tx, classified_data->port);
+				(uint16_t)(classified_data->num_pkt - n_tx),
+				classified_data->port);
 	}
 
 	classified_data->num_pkt = 0;
